changed tempValorDia to parse valorDia.textBox.Text to float instead of int

+++ Components/Forms/CarroForm.cs
@@ -234,7 +234,7 @@ namespace Rent_a_Car.Components.Forms
                     }
                     try
                     {
-                        tempValorDia = Int32.Parse(valorDia.textBox.Text);
+                        tempValorDia = float.Parse(valorDia.textBox.Text);
                     }
                     catch (Exception ex)
                     {
